# backend/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Finance Tracker API
  description: API for tracking personal finances with JWT authentication
  version: 1.0.0
tags:
  - name: Authentication
    description: User authentication and registration
  - name: User
    description: User profile operations
  - name: Transactions
    description: Operations about financial transactions
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  # Authentication endpoints
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      security: []  # Override global security - no auth required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Registration failed
          
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      operationId: loginUser
      security: []  # Override global security - no auth required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials
          
  # User endpoints
  /api/user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized

  /api/transactions:
    get:
      tags:
        - Transactions
      summary: Get all transactions
      operationId: getAllTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
    
    post:
      tags:
        - Transactions
      summary: Create a new transaction
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      operationId: getTransactionById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized
        '404':
          description: Transaction not found
        '500':
          description: Internal server error
    
    put:
      tags:
        - Transactions
      summary: Update transaction
      operationId: updateTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '200':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          description: Unauthorized
        '404':
          description: Transaction not found
        '500':
          description: Internal server error
    
    delete:
      tags:
        - Transactions
      summary: Delete transaction
      operationId: deleteTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Transaction deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Transaction not found
        '500':
          description: Internal server error

components:
  schemas:
    Transaction:
      type: object
      required:
        - amount
        - type
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        amount:
          type: number
          format: double
          description: Transaction amount
        type:
          type: string
          enum: [INCOME, EXPENSE]
          description: Type of transaction
        category:
          type: string
          description: Transaction category
        description:
          type: string
          description: Transaction description
        dateTime:
          type: string
          format: date-time
          description: Transaction date and time

    # Authentication DTOs
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Unique username
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password
          
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username
        password:
          type: string
          description: User password
          
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        type:
          type: string
          default: "Bearer"
          description: Token type
        username:
          type: string
          description: Username
        email:
          type: string
          description: User email
          
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          description: User email
        role:
          type: string
          enum: [USER, ADMIN]
          description: User role
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Apply security to all transaction endpoints
security:
  - bearerAuth: []